% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews.R and run devtools::document()
\name{get_reviews}
\alias{get_reviews}
\title{Easily Web Scrapes Glassdoor Company Reviews Into a Data Frame}
\usage{
get_reviews(companyNum)
}
\arguments{
\item{companyNum}{A string representing a company's unique ID number. Identified by navigating to a company's Glassdoor reviews web page
and reviewing the URL for characters between "Reviews-" and ".htm" (usually starts with an 'E' and followed by up to seven digits).}
}
\value{
\code{get_reviews} returns a data frame containing reviews and source information.
}
\description{
Generate a data frame of company reviews with one function: \code{get_reviews()}.
}
\examples{
Reference https://www.glassdoor.com/Reviews/SpaceX-Reviews-E40371.htm
String enclosed with quotes.
  reviews <- get_reviews(companyNum = "E40371")
  reviews <- get_reviews("E40371")

}
